ctdb: make enabling ctdb hook scripts configurable
It is not always desirable to have the ctdb hook scripts start/stop
the ctdb service and fiddle with the fstab for the lock volume.
This patch introduces a variable
gluster_features_ctdb_enable_hook_script defaulting to true which
can be set to false to disable the enablement of the hook scripts.

Signed-off-by: Michael Adam <obnox@samba.org>